Move unit tests to the correct package for class they are testing
